Using commits since last tag as the PATCH version.
Gradle versioning uses lexical ordering of the qualifier so it considers "1.2.3-something" less than "1.2.3".

This means that gradle treats unreleased versions as being less than released versions.

This uses `commitsSinceLastTag` as the patch version so each additional commit on a tag should naturally be seen as greater than the previous commits